refactor: split top page into multiple indepedent components
