+++ client/.eslintrc.json
@@ -24,5 +24,7 @@
       }
     }
   ],
-  "rules": {}
+  "rules": {
+    "react/no-unknown-property": ["error", { "ignore": ["css"] }]
+  }
 }
