+++ client/components/model/product/Product.tsx
@@ -1,40 +1,6 @@
 import { Flex, Box, Image, Badge, useColorModeValue } from '@chakra-ui/react';
 import { BsStar, BsStarFill, BsStarHalf } from 'react-icons/bs';
-import { currencyFormat } from '../../utils/format';
-
-interface RatingProps {
-  rating: number;
-  numReviews: number;
-}
-
-function Rating({ rating, numReviews }: RatingProps) {
-  return (
-    <Box display="flex" alignItems="center">
-      {Array(5)
-        .fill('')
-        .map((_, i) => {
-          const roundedRating = Math.round(rating * 2) / 2;
-          if (roundedRating - i >= 1) {
-            return (
-              <BsStarFill
-                key={i}
-                style={{ marginLeft: '1' }}
-                color={i < rating ? 'teal.500' : 'gray.300'}
-              />
-            );
-          }
-          if (roundedRating - i === 0.5) {
-            return <BsStarHalf key={i} style={{ marginLeft: '1' }} />;
-          }
-          return <BsStar key={i} style={{ marginLeft: '1' }} />;
-        })}
-      <Box as="span" ml="2" color="gray.600" fontSize="sm">
-        ({numReviews})
-      </Box>
-    </Box>
-  );
-}
-
+import { currencyFormat } from '../../function/utils/format';
 interface ProductProps {
   isNewItem: boolean;
   imageURL: string;
@@ -45,7 +11,8 @@ interface ProductProps {
   selected: boolean;
   handleClick: () => void;
 }
-function Product({
+
+const Product = ({
   isNewItem,
   imageURL,
   name,
@@ -54,7 +21,7 @@ function Product({
   numReviews,
   selected,
   handleClick,
-}: ProductProps) {
+}: ProductProps) => {
   return (
     <Flex p="2" alignItems="center" justifyContent="center">
       <Box
@@ -99,6 +66,39 @@ function Product({
         </Box>
       </Box>
     </Flex>
+  );
+};
+
+interface RatingProps {
+  rating: number;
+  numReviews: number;
+}
+
+function Rating({ rating, numReviews }: RatingProps) {
+  return (
+    <Box display="flex" alignItems="center">
+      {Array(5)
+        .fill('')
+        .map((_, i) => {
+          const roundedRating = Math.round(rating * 2) / 2;
+          if (roundedRating - i >= 1) {
+            return (
+              <BsStarFill
+                key={i}
+                style={{ marginLeft: '1' }}
+                color={i < rating ? 'teal.500' : 'gray.300'}
+              />
+            );
+          }
+          if (roundedRating - i === 0.5) {
+            return <BsStarHalf key={i} style={{ marginLeft: '1' }} />;
+          }
+          return <BsStar key={i} style={{ marginLeft: '1' }} />;
+        })}
+      <Box as="span" ml="2" color="gray.600" fontSize="sm">
+        ({numReviews})
+      </Box>
+    </Box>
   );
 }
