+++ client/components/shared/invoice/index.tsx
@@ -1,0 +1,31 @@
+export { default as SimpleInvoice } from './SimpleInvoice';
+export { default as DetailInvoice } from './DetailInvoice';
+
+export type TInvoice = {
+  id: string;
+  label: string;
+  details: {
+    id: string;
+    label: string;
+    value: number;
+  }[];
+  total: {
+    label: string;
+    value: number;
+  };
+};
+
+export const defaultInvoiceTable: TInvoice[] = [
+  {
+    id: '1',
+    label: 'Monthly Fees',
+    details: [],
+    total: { label: 'Total', value: 0 },
+  },
+  {
+    id: '2',
+    label: 'One-Time Fees',
+    details: [],
+    total: { label: 'Total', value: 0 },
+  },
+];
