+++ client/components/types/product.ts
